# packages/contracts/openapi/careos-api.v1.json
{
  "components": {
    "schemas": {
...
      "ConnectionStatus": {
        "enum": [
          "ONLINE",
          "OFFLINE",
          "UNKNOWN"
        ],
        "title": "ConnectionStatus",
        "type": "string"
      },
      "ConnectionView": {
        "properties": {
          "battery_level": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Battery Level"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          },
          "signal_strength": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal Strength"
          },
          "status": {
            "$ref": "#/components/schemas/ConnectionStatus"
          }
        },
        "required": [
          "status",
          "last_seen_at",
          "battery_level",
          "signal_strength"
        ],
        "title": "ConnectionView",
        "type": "object"
      },
      "CreateDeviceRequest": {
        "properties": {
          "adapter": {
            "default": "careos",
            "pattern": "^[a-z0-9_\\-]{1,40}$",
            "title": "Adapter",
            "type": "string"
          },
          "device_type": {
            "$ref": "#/components/schemas/DeviceType"
          },
          "external_id": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9_\\-:.]*$",
            "title": "External Id",
            "type": "string"
          },
          "manufacturer": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Manufacturer",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "service_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service User Id"
          }
        },
        "required": [
          "external_id",
          "device_type",
          "manufacturer"
        ],
        "title": "CreateDeviceRequest",
        "type": "object"
      },
...
      "DeviceStatus": {
        "enum": [
          "ACTIVE",
          "INACTIVE",
          "RETIRED"
        ],
        "title": "DeviceStatus",
        "type": "string"
      },
      "DeviceType": {
        "enum": [
          "SOS_PENDANT",
          "SMARTWATCH",
          "FALL_SENSOR",
          "HOME_HUB",
          "OTHER"
        ],
        "title": "DeviceType",
        "type": "string"
      },
      "DeviceView": {
        "properties": {
          "adapter": {
            "title": "Adapter",
            "type": "string"
          },
          "connection": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConnectionView"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "device_type": {
            "$ref": "#/components/schemas/DeviceType"
          },
          "external_id": {
            "title": "External Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "low_battery": {
            "title": "Low Battery",
            "type": "boolean"
          },
          "manufacturer": {
            "title": "Manufacturer",
            "type": "string"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "service_user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ServiceUserRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/DeviceStatus"
          }
        },
        "required": [
          "id",
          "external_id",
          "device_type",
          "manufacturer",
          "model",
          "adapter",
          "status",
          "service_user",
          "connection",
          "low_battery",
          "created_at"
        ],
        "title": "DeviceView",
        "type": "object"
      },
      "EscalationActionType": {
        "enum": [
          "AUTOMATED_USER_CONTACT",
          "CALL_TRUSTED_CONTACT",
          "NOTIFY_TRUSTED_CONTACT",
          "OPERATOR_ESCALATION"
        ],
        "title": "EscalationActionType",
        "type": "string"
      },
      "EscalationPolicyView": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "is_default": {
            "title": "Is Default",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/EscalationStepView"
            },
            "title": "Steps",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "is_default",
          "is_active",
          "revision",
          "steps",
          "updated_at"
        ],
        "title": "EscalationPolicyView",
        "type": "object"
      },
      "EscalationPolicyWrite": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_active": {
            "default": true,
            "title": "Is Active",
            "type": "boolean"
          },
          "is_default": {
            "default": false,
            "title": "Is Default",
            "type": "boolean"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/EscalationStepWrite"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "name",
          "steps"
        ],
        "title": "EscalationPolicyWrite",
        "type": "object"
      },
      "EscalationStepView": {
        "properties": {
          "action_type": {
            "$ref": "#/components/schemas/EscalationActionType"
          },
          "contact_priority": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Priority"
          },
          "delay_seconds": {
            "title": "Delay Seconds",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "step_order": {
            "title": "Step Order",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "step_order",
          "delay_seconds",
          "action_type",
          "contact_priority"
        ],
        "title": "EscalationStepView",
        "type": "object"
      },
      "EscalationStepWrite": {
        "properties": {
          "action_type": {
            "$ref": "#/components/schemas/EscalationActionType"
          },
          "contact_priority": {
            "anyOf": [
              {
                "maximum": 9.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Priority"
          },
          "delay_seconds": {
            "maximum": 86400.0,
            "minimum": 0.0,
            "title": "Delay Seconds",
            "type": "integer"
          }
        },
        "required": [
          "delay_seconds",
          "action_type"
        ],
        "title": "EscalationStepWrite",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngestResponse": {
        "properties": {
          "duplicate": {
            "title": "Duplicate",
            "type": "boolean"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "incident_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident Id"
          },
          "incident_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident Reference"
          },
          "incident_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident Status"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReceiptOutcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "receipt_id": {
            "format": "uuid",
            "title": "Receipt Id",
            "type": "string"
          }
        },
        "required": [
          "receipt_id",
          "event_id",
          "duplicate",
          "outcome",
          "incident_id",
          "incident_reference",
          "incident_status"
        ],
        "title": "IngestResponse",
        "type": "object"
      },
...
      "ReceiptOutcome": {
        "enum": [
          "INCIDENT_CREATED",
          "ATTACHED_TO_INCIDENT",
          "TELEMETRY_ONLY"
        ],
        "title": "ReceiptOutcome",
        "type": "string"
      },
...
      "ServiceUserRef": {
        "properties": {
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name",
          "city"
        ],
        "title": "ServiceUserRef",
        "type": "object"
      },
...
      "SimulatorEventRequest": {
        "properties": {
          "battery": {
            "default": 84,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Battery",
            "type": "integer"
          },
          "event_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 8,
                "pattern": "^[A-Za-z0-9][A-Za-z0-9_\\-:.]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Reuse a previous event_id to demonstrate gateway idempotency.",
            "title": "Event Id"
          },
          "event_type": {
            "default": "SOS_BUTTON",
            "enum": [
              "SOS_BUTTON",
              "FALL_DETECTED",
              "DEVICE_FAULT",
              "LOW_BATTERY",
              "HEARTBEAT"
            ],
            "title": "Event Type",
            "type": "string"
          },
          "latitude": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "signal": {
            "default": 92,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Signal",
            "type": "integer"
          },
          "timestamp_ms": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Device timestamp. Send the original value with event_id for an exact replay.",
            "title": "Timestamp Ms"
          }
        },
        "title": "SimulatorEventRequest",
        "type": "object"
      },
      "SimulatorEventResult": {
        "properties": {
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "gateway_response": {
            "additionalProperties": true,
            "title": "Gateway Response",
            "type": "object"
          },
          "gateway_status_code": {
            "title": "Gateway Status Code",
            "type": "integer"
          },
          "vendor_payload": {
            "additionalProperties": true,
            "title": "Vendor Payload",
            "type": "object"
          }
        },
        "required": [
          "event_id",
          "vendor_payload",
          "gateway_status_code",
          "gateway_response"
        ],
        "title": "SimulatorEventResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "CSRF": {
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "GatewayKey": {
        "in": "header",
        "name": "X-CareOS-Gateway-Key",
        "type": "apiKey"
      },
      "Session": {
        "in": "cookie",
        "name": "careos_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "CareOS does not diagnose, predict illness or recommend treatment. AI features are assistive only; incident handling is deterministic.",
    "summary": "Telecare & safety incident orchestration platform",
    "title": "CareOS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/devices": {
      "get": {
        "operationId": "list_devices_v1_devices_get",
        "parameters": [
          {
            "in": "query",
            "name": "service_user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Service User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DeviceView"
                  },
                  "title": "Response List Devices V1 Devices Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "List Devices",
        "tags": [
          "devices"
        ]
      },
      "post": {
        "operationId": "create_device_v1_devices_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "Create Device",
        "tags": [
          "devices"
        ]
      }
    },
    "/v1/escalation-policies": {
      "get": {
        "operationId": "list_policies_v1_escalation_policies_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EscalationPolicyView"
                  },
                  "title": "Response List Policies V1 Escalation Policies Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "List Policies",
        "tags": [
          "escalation policies"
        ]
      },
      "post": {
        "operationId": "create_policy_v1_escalation_policies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EscalationPolicyWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EscalationPolicyView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "Create Policy",
        "tags": [
          "escalation policies"
        ]
      }
    },
    "/v1/escalation-policies/{policy_id}": {
      "put": {
        "operationId": "replace_policy_v1_escalation_policies__policy_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "policy_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Policy Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EscalationPolicyWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EscalationPolicyView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "Replace Policy",
        "tags": [
          "escalation policies"
        ]
      }
    },
    "/v1/gateway/{adapter}/events": {
      "post": {
        "description": "Accepts a vendor payload for the named adapter. The `careos` adapter accepts the CareOS Device Event Contract directly (JSON Schema: packages/contracts/schemas/careos-device-event.v1.schema.json). Idempotent on `event_id`: redelivery returns the original receipt with `duplicate=true` and never creates a second incident.",
        "operationId": "ingest_event_v1_gateway__adapter__events_post",
        "parameters": [
          {
            "in": "path",
            "name": "adapter",
            "required": true,
            "schema": {
              "pattern": "^[a-z0-9_\\-]{1,40}$",
              "title": "Adapter",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "GatewayKey": []
          }
        ],
        "summary": "Receive a device event",
        "tags": [
          "device-gateway"
        ]
      }
    },
...
    "/v1/simulator/devices/{device_id}/events": {
      "post": {
        "operationId": "simulate_device_event_v1_simulator_devices__device_id__events_post",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulatorEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulatorEventResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "Session": []
          },
          {
            "CSRF": []
          }
        ],
        "summary": "Simulate Device Event",
        "tags": [
          "simulator (development only)"
        ]
      }
    },
...
  }
}